backend/myproject/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.7.14
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-freeze==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
iniconfig==2.1.0
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xmltodict==0.14.2
psutil==5.9.5
# Pillow==10.0.1
# channels==4.0.0
channels-redis==4.1.0
# daphne==4.0.0